requirements.txt
Click==7.0
Flask==1.1.2
Jinja2==2.11.3
python-dotenv==1.2.1